Adding comparison between builds
Adds the feature to compare diffs between builds to the plugin based on the builds's SHA and build numbers.

The new feature is the only one that can evaluate to "true" to be used.

The feature currently requires jq to be ran successfully.

The feature requires API access key with `read_builds`permissions.

WIP at the moment.